feat: organize test files into proper test/ directories
Core package improvements:
- Move all *.test.ts files from src/ to test/ directory
- Fix import paths to reference ../src/ from test files
- Create vitest.config.ts for proper test configuration
- Fix typo: 'vitest runce' → 'vitest run' in package.json
